requirements.txt
aioboto3
aiofiles
aiohttp
click
python-dotenv
tqdm